results/rebuttal_new_baselines/run-20231112-235354_LLMatic-ZeroShot_donnemartin-system-design-oop-finance2_0_20-runs_log/finance2/3/LLMatic/models/budget.py
class Budget:
	def __init__(self, user, amount, category):
		self.user = user
		self.amount = amount
		self.category = category

	@staticmethod
	def set_budget(user, amount, category):
		# Mocking setting budget
		return {'user': user, 'amount': amount, 'category': category}

	@staticmethod
	def adjust_budget(user, amount):
		# Mocking adjusting budget
		return {'user': user, 'amount': amount}

	@staticmethod
	def get_user_budgets(user):
		# Mocking getting user budgets
		return [{'user': user, 'amount': 500, 'category': 'Groceries'}]
